由夜零发起的VPC SDK自动发布, BUILD_ID=145, 版本号：3.0.1 发布日志： 1, 修改了一些方法的访问次数。 2, 新增了错误码。

+++ aliyun-java-sdk-vpc/src/main/java/com/aliyuncs/vpc/model/v20160428/CreateBgpGroupRequest.java
@@ -44,7 +44,7 @@ public class CreateBgpGroupRequest extends RpcAcsRequest<CreateBgpGroupResponse>
 
 	private Long ownerId;
 
-	private Integer peerAsn;
+	private Long peerAsn;
 
 	private Boolean isFakeAsn;
 
@@ -129,11 +129,11 @@ public class CreateBgpGroupRequest extends RpcAcsRequest<CreateBgpGroupResponse>
 		}
 	}
 
-	public Integer getPeerAsn() {
+	public Long getPeerAsn() {
 		return this.peerAsn;
 	}
 
-	public void setPeerAsn(Integer peerAsn) {
+	public void setPeerAsn(Long peerAsn) {
 		this.peerAsn = peerAsn;
 		if(peerAsn != null){
 			putQueryParameter("PeerAsn", peerAsn.toString());
